refactor serializer, onehot fxns, added almost working tensorflow model on all games
