+++ sips/macros/nfl.py
@@ -4,7 +4,7 @@ teams = ['Arizona Cardinals',
 'Buffalo Bills',
 'Carolina Panthers',
 'Chicago Bears',
-'Cincinnati Bengals',
+'Cincinnati Bengals ',
 'Cleveland Browns',
 'Dallas Cowboys',
 'Denver Broncos',
