+++ sips/macros/nhl.py
@@ -22,7 +22,7 @@ teams = ['Anaheim Ducks',
 'Philadelphia Flyers',
 'Pittsburgh Penguins',
 'San Jose Sharks',
-'St Louis Blues',
+'St. Louis Blues',
 'Tampa Bay Lightning',
 'Toronto Maple Leafs',
 'Vancouver Canucks',
